+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
 ## [v0.0.1]
-
 - Initial release
+
+## [v0.0.2]
+- Path Intellisense support
+- Resources Intellisense support
+- Build project/module support